Change Auth1/Auth2 clients by one single class
