--- src/ohmy/Support/Traits/OverloadMethod.php
@@ -1,30 +1,0 @@
-<?php namespace ohmy\Support\Traits;
-
-trait OverloadMethod
-{
-    /**
-     * Handle dynamic method calls into the method.
-     *
-     * @param  string $method
-     * @param  array  $parameters
-     * @return mixed
-     */
-    public function __call($method, $parameters)
-    {
-        return call_user_func_array(array($this, $method), $parameters);
-    }
-
-    /**
-     * Handle dynamic static method calls into the method.
-     *
-     * @param  string $method
-     * @param  array  $parameters
-     * @return Mixed
-     */
-    public static function __callStatic($method, $parameters)
-    {
-        $instance = new static;
-
-        return call_user_func_array(array($instance, $method), $parameters);
-    }
-}
